requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
Django==4.2.7
django-debug-toolbar==4.2.0
django-elasticsearch-debug-toolbar==3.0.2
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-nine==0.2.7
djangorestframework==3.14.0
drf-extensions==0.7.1
elastic-transport==8.10.0
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
Faker==19.13.0
importlib-metadata==6.8.0
Markdown==3.5.1
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
zipp==3.17.0
